+++ services/tunnelbroker/src/Tools/Tools.h
@@ -8,7 +8,7 @@ namespace network {
 namespace tools {
 
 std::string generateRandomString(std::size_t length);
-long long getCurrentTimestamp();
+int64_t getCurrentTimestamp();
 bool validateDeviceID(std::string deviceID);
 std::string generateUUID();
 bool validateSessionID(std::string sessionID);
